added groups as a new application
