+++ mysite/chat/models.py
@@ -42,11 +42,3 @@ class Message(models.Model):
     @property
     def is_updated(self):
         return self.sent_at != self.updated_at
-
-
-class Group(models.Model):
-    group_name = models.CharField(max_length=50, null=False)
-    group_description = models.CharField(null=True)
-    owner = models.CharField(max_length=50)
-
-    users = models.ManyToManyField(get_user_model())
