+++ mysite/chat/serializers.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
 from authorization.models import User
-from chat.models import Message, Profile, Group, PrivateChat, PublicChat
+from chat.models import Message, Profile, PrivateChat, PublicChat
 
 
 class PublicChatSerializer(serializers.ModelSerializer):
@@ -21,12 +21,6 @@ class ChatSerializer(serializers.Serializer):
     created_at = serializers.DateTimeField(read_only=True)
     chat_name = serializers.CharField(read_only=True, allow_blank=True, required=False, allow_null=True)
     users = serializers.PrimaryKeyRelatedField(queryset=User.objects.all(), many=True)
-
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ('pk', 'username', 'is_premium', 'is_active')
 
 
 class ProfileSerializer(serializers.ModelSerializer):
@@ -61,12 +55,3 @@ class ProfileDetailSerializer(serializers.ModelSerializer):
     class Meta:
         model = Profile
         fields = ('user_id', 'user_info', 'first_name', 'last_name', 'created_at')
-
-
-class GroupSerializer(serializers.ModelSerializer):
-    users = UserSerializer(many=True, read_only=True)
-
-    class Meta:
-        model = Group
-        fields = ('id', 'group_name', "owner", 'group_description', "users")
-        read_only_fields = ['owner']
