+++ mysite/chat/views.py
@@ -1,12 +1,10 @@
-import rest_framework.views
 from rest_framework import generics, mixins, permissions
 from rest_framework.exceptions import ValidationError, PermissionDenied
-from rest_framework.response import Response
-from chat.permissions import DjangoModelPermissionsWithRead, IsChatMember, DjangoObjectPermissionsWithReadUpdate
-from chat.filters import GroupFilter, MessageFilter
-from chat.models import Message, Profile, Group, PrivateChat, PublicChat
+from chat.permissions import IsChatMember
+from chat.filters import MessageFilter
+from chat.models import Message, Profile, PrivateChat, PublicChat
 from chat.serializers import PrivateMessageSerializer, ProfileSerializer, \
-    ProfileDetailSerializer, GroupSerializer, PrivateChatSerializer, PublicChatSerializer, PublicMessageSerializer, \
+    ProfileDetailSerializer, PrivateChatSerializer, PublicChatSerializer, PublicMessageSerializer, \
     ChatSerializer
 
 
@@ -133,8 +131,31 @@ class CreateUpdateGetDeletePublicMessageView(
     def perform_create(self, serializer):
         public_chat = self.kwargs.get('public_chat_id')
 
+        user = self.request.user
+
+        if user not in public_chat.users.all():
+            raise PermissionDenied("You do not have permission to perform this action")
+
         serializer.save(user=self.request.user, public_chat_id=public_chat)
 
+    def perform_update(self, serializer):
+        public_chat = self.kwargs.get('public_chat_id')
+        user = self.request.user
+
+        if user not in public_chat.users.all():
+            raise PermissionDenied("You do not have permission to perform this action")
+
+        serializer.save(user=self.request.user, public_chat_id=public_chat)
+
+    def perform_destroy(self, serializer):
+
+        public_chat = self.kwargs.get('public_chat_id')
+        user = self.request.user
+
+        if user not in public_chat.users.all():
+            raise PermissionDenied("You do not have permission to perform this action")
+
+        serializer.save(user=self.request.user, public_chat_id=public_chat)
 
 class CreateUpdateGetDeletePrivateMessageView(
     generics.GenericAPIView,
@@ -185,7 +206,6 @@ class CurrentUserProfileView(generics.GenericAPIView,
                              mixins.UpdateModelMixin,
                              mixins.DestroyModelMixin,
                              mixins.RetrieveModelMixin):
-
     queryset = Profile.objects.all()
     serializer_class = ProfileSerializer
 
@@ -205,73 +225,9 @@ class CurrentUserProfileView(generics.GenericAPIView,
         return instance
 
 
-
 class RetrieveProfileView(generics.RetrieveAPIView):
     queryset = Profile.objects.all()
     serializer_class = ProfileDetailSerializer
     lookup_field = 'user_id'
 
     permissions_classes = [permissions.DjangoModelPermissions]
-
-
-class CreateGroupView(generics.CreateAPIView):
-    queryset = Group.objects.all()
-    serializer_class = GroupSerializer
-    lookup_field = 'pk'
-
-    permission_classes = [permissions.DjangoModelPermissions]
-
-    def perform_create(self, serializer):
-        serializer.save(owner=self.request.user)
-
-
-class GroupsListView(generics.GenericAPIView, mixins.ListModelMixin, mixins.DestroyModelMixin, mixins.UpdateModelMixin):
-    filterset_class = GroupFilter
-
-    queryset = Group.objects.all()
-    serializer_class = GroupSerializer
-    lookup_field = 'pk'
-
-    permission_classes = [DjangoModelPermissionsWithRead]
-
-
-class GroupsDetailView(generics.RetrieveUpdateAPIView):
-    queryset = Group.objects.all()
-    serializer_class = GroupSerializer
-    lookup_field = 'pk'
-    permission_classes = [DjangoObjectPermissionsWithReadUpdate]
-
-    def get(self, request, *args, **kwargs):
-        return self.retrieve(request, *args, **kwargs)
-
-    def patch(self, request, *args, **kwargs):
-        return self.partial_update(request, *args, **kwargs)
-
-    def get_queryset(self):
-        group = self.kwargs.get('pk')
-        qs = Group.objects.all().filter(pk=group)
-        return qs
-
-
-class FollowGroupView(rest_framework.views.APIView):
-    queryset = Group.objects.all()
-    lookup_field = 'pk'
-
-    def post(self, request, *args, **kwargs):
-        group_id = kwargs.get('pk')
-        request.user.group_set.add(group_id)
-        return Response({'success followed': 'True'})
-
-
-class UnfollowGroupView(rest_framework.views.APIView):
-    queryset = Group.objects.all()
-    lookup_field = 'pk'
-
-    permission_classes = [permissions.DjangoModelPermissions]
-
-    http_method_names = ['delete']
-
-    def delete(self, request, *args, **kwargs):
-        group_id = kwargs.get('pk')
-        request.user.group_set.remove(group_id)
-        return Response({'success unfollowed': 'True'})
