+++ mysite/mysite/urls.py
@@ -25,6 +25,7 @@ Including another URLconf
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('chat/', include('chat.urls')),
+    path('group/', include('group.urls')),
     path('auth/', include('authorization.urls'))
 ]
